refactor(page-manager,view): add generic method to add data to pageManager
Variable assignment was to loose, so I add a public method called 'addData()'
which add a key/value pair to the pageManager's internal $data array.
This allow standardization and readibility to view elaboration. Moreover,
view's information are't "fetch" from it anymore, but added to it (all
views are born egal).

Signed-off-by: Édouard Lopez <dev+git@edouard-lopez.com>